feat: Implement dark mode feature
This commit introduces a dark mode feature to the Phialo Design project.

Key changes include:
- Configured Tailwind CSS to use the 'class' dark mode strategy.
- Added a FOUC prevention script to BaseLayout.astro to ensure smooth theme transitions.
- Created a reusable ThemeToggle.astro component for switching between light and dark modes.
- Integrated the ThemeToggle component into the Header.astro.
- Updated styling in BaseLayout.astro, Header.astro, and Footer.astro to support dark mode using Tailwind's dark: variants.
- Re-enabled and updated E2E tests in integration.spec.ts and responsive.spec.ts to cover dark mode functionality and persistence.
- Removed the temporary dark-mode.spec.ts test file.

Resolves #42
