+++ phialo-design/src/shared/components/ui/ThemeToggle.astro
@@ -1,0 +1,32 @@
+---
+// src/components/ThemeToggle.astro
+---
+<button id="theme-toggle" aria-label="Toggle dark mode">
+  <!-- You can use icons for sun and moon here -->
+  <span class="light-icon">☀️</span>
+  <span class="dark-icon">🌙</span>
+</button>
+
+<style>
+  .dark .light-icon {
+    display: none;
+  }
+  .dark-icon {
+    display: none;
+  }
+  .dark .dark-icon {
+    display: block;
+  }
+</style>
+
+<script>
+  document.addEventListener('DOMContentLoaded', () => {
+    const toggle = document.getElementById('theme-toggle');
+    if (toggle) {
+      toggle.addEventListener('click', () => {
+        const isDark = document.documentElement.classList.toggle('dark');
+        localStorage.setItem('theme', isDark ? 'dark' : 'light');
+      });
+    }
+  });
+</script>
