+++ phialo-design/src/shared/layouts/BaseLayout.astro
@@ -40,6 +40,23 @@ const englishUrl = new URL('/en' + basePath, Astro.site).href;
 <!DOCTYPE html>
 <html lang={lang} class="scroll-smooth">
   <head>
+    <script is:inline>
+      const theme = (() => {
+        if (typeof localStorage !== 'undefined' && localStorage.getItem('theme')) {
+          return localStorage.getItem('theme');
+        }
+        if (window.matchMedia('(prefers-color-scheme: dark)').matches) {
+          return 'dark';
+        }
+        return 'light';
+      })();
+
+      if (theme === 'dark') {
+        document.documentElement.classList.add('dark');
+      } else {
+        document.documentElement.classList.remove('dark');
+      }
+    </script>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="generator" content={Astro.generator} />
@@ -97,7 +114,8 @@ const englishUrl = new URL('/en' + basePath, Astro.site).href;
     />
     
     <ViewTransitions />
-  </head>  <body class="min-h-screen antialiased">
+  </head>
+  <body class="min-h-screen antialiased bg-white text-black dark:bg-gray-900 dark:text-white theme-transition">
     <SmoothScroll client:load />
     <slot />
