+++ phialo-design/tests/e2e/integration.spec.ts
@@ -53,6 +53,97 @@ test.describe('Integration Tests - All Fixes', () => {
     // Dark mode persistence test moved to tests/future/dark-mode.spec.ts
   });
 
+  test('Should toggle between light and dark modes', async ({ page }) => {
+    await page.goto('/');
+    
+    // Find and click theme toggle button
+    const themeToggle = page.locator('#theme-toggle');
+    await expect(themeToggle).toBeVisible();
+    
+    // Click to enable dark mode
+    await themeToggle.click();
+    await expect(page.locator('html')).toHaveClass(/dark/);
+    
+    // Verify dark mode styles are applied
+    const backgroundColor = await page.locator('body').evaluate(el => 
+      window.getComputedStyle(el).backgroundColor
+    );
+    expect(backgroundColor).toBe('rgb(17, 24, 39)'); // Corresponds to dark:bg-gray-900
+    
+    // Click to return to light mode
+    await themeToggle.click();
+    await expect(page.locator('html')).not.toHaveClass(/dark/);
+  });
+
+  test('Should persist theme preference across page navigation', async ({ page }) => {
+    await page.goto('/');
+    
+    // Enable dark mode
+    const themeToggle = page.locator('#theme-toggle');
+    await themeToggle.click();
+    await expect(page.locator('html')).toHaveClass(/dark/);
+    
+    // Navigate to another page
+    await page.goto('/portfolio');
+    
+    // Theme should still be dark
+    await expect(page.locator('html')).toHaveClass(/dark/);
+  });
+
+  test('Language switching preserves dark mode', async ({ page }) => {
+    await page.goto('/');
+    
+    // Enable dark mode
+    const themeToggle = page.locator('#theme-toggle');
+    await themeToggle.click();
+    await expect(page.locator('html')).toHaveClass(/dark/);
+    
+    // Switch language
+    const englishButton = page.locator('[data-language-selector] a[href="/en/"]');
+    await englishButton.click();
+    await expect(page).toHaveURL('/en/');
+    
+    // Dark mode should be preserved
+    await expect(page.locator('html')).toHaveClass(/dark/);
+  });
+
+  test('Should respect system preference on first visit', async ({ page }) => {
+    // Set browser to prefer dark mode
+    await page.emulateMedia({ colorScheme: 'dark' });
+    
+    await page.goto('/');
+    
+    // Should automatically be in dark mode
+    await expect(page.locator('html')).toHaveClass(/dark/);
+    
+    // Set browser to prefer light mode
+    await page.emulateMedia({ colorScheme: 'light' });
+    await page.reload();
+    
+    // Should automatically be in light mode
+    await expect(page.locator('html')).not.toHaveClass(/dark/);
+  });
+
+  test('Theme transition should be smooth', async ({ page }) => {
+    await page.goto('/');
+    
+    // Add listener for transition events
+    const transitionPromise = page.evaluate(() => {
+      return new Promise(resolve => {
+        document.body.addEventListener('transitionend', () => resolve(true), { once: true });
+        setTimeout(() => resolve(false), 1000); // Timeout after 1 second
+      });
+    });
+    
+    // Toggle theme
+    const themeToggle = page.locator('#theme-toggle');
+    await themeToggle.click();
+    
+    // Verify transition occurred
+    const hadTransition = await transitionPromise;
+    expect(hadTransition).toBe(true);
+  });
+
   test('@critical Complete user journey in German (CI-friendly)', async ({ page }) => {
     // Skip this test entirely in CI - too many navigation issues
     test.skip(!!process.env.CI, 'Navigation tests are unreliable in CI environment');
