--- phialo-design/tests/future/dark-mode.spec.ts
@@ -1,143 +1,0 @@
-import { test, expect } from '@playwright/test';
-
-/**
- * Dark Mode / Theme Toggle Tests
- * 
- * These tests are pending implementation of the dark mode feature (Issue #12).
- * Once the feature is implemented, move these tests back to the e2e directory
- * and remove the skip annotations.
- * 
- * Implementation checklist for dark mode:
- * - [ ] Add theme toggle button to header/navigation
- * - [ ] Implement theme switching logic with localStorage persistence
- * - [ ] Add CSS variables for light/dark themes
- * - [ ] Ensure theme persists across page navigation
- * - [ ] Add proper ARIA labels for accessibility
- * - [ ] Test on all supported browsers and devices
- */
-
-test.describe('Dark Mode Tests (Future Feature)', () => {
-  test.skip('Should toggle between light and dark modes', async ({ page }) => {
-    await page.goto('/');
-    
-    // Find and click theme toggle button
-    const themeToggle = page.locator('button[aria-label*="mode"]');
-    await expect(themeToggle).toBeVisible();
-    
-    // Click to enable dark mode
-    await themeToggle.click();
-    await expect(page.locator('html')).toHaveClass(/theme-dark/);
-    
-    // Verify dark mode styles are applied
-    const backgroundColor = await page.locator('body').evaluate(el => 
-      window.getComputedStyle(el).backgroundColor
-    );
-    expect(backgroundColor).not.toBe('rgb(255, 255, 255)'); // Not white
-    
-    // Click to return to light mode
-    await themeToggle.click();
-    await expect(page.locator('html')).not.toHaveClass(/theme-dark/);
-  });
-
-  test.skip('Should persist theme preference across page navigation', async ({ page }) => {
-    await page.goto('/');
-    
-    // Enable dark mode
-    const themeToggle = page.locator('button[aria-label*="mode"]');
-    await themeToggle.click();
-    await expect(page.locator('html')).toHaveClass(/theme-dark/);
-    
-    // Navigate to another page
-    await page.goto('/portfolio');
-    
-    // Theme should still be dark
-    await expect(page.locator('html')).toHaveClass(/theme-dark/);
-    
-    // Verify toggle button shows correct state
-    const toggleAriaLabel = await themeToggle.getAttribute('aria-label');
-    expect(toggleAriaLabel).toContain('light'); // Should show option to switch to light
-  });
-
-  test.skip('Language switching preserves dark mode', async ({ page }) => {
-    await page.goto('/');
-    
-    // Enable dark mode
-    const themeToggle = page.locator('button[aria-label*="mode"]');
-    await themeToggle.click();
-    await expect(page.locator('html')).toHaveClass(/theme-dark/);
-    
-    // Switch language
-    const englishButton = page.locator('[data-language-selector] button[data-lang="en"]');
-    await englishButton.click();
-    await expect(page).toHaveURL('/en/');
-    
-    // Dark mode should be preserved
-    await expect(page.locator('html')).toHaveClass(/theme-dark/);
-  });
-
-  test.skip('@critical Theme toggle should work on mobile @mobile', async ({ page, isMobile }) => {
-    test.skip(!isMobile, 'Mobile only test');
-    
-    await page.goto('/');
-    
-    // Open mobile menu if needed
-    const menuButton = page.locator('[data-testid="mobile-menu-button"]');
-    if (await menuButton.isVisible()) {
-      await menuButton.click();
-      await page.waitForSelector('[data-testid="mobile-menu-panel"]', { state: 'visible' });
-    }
-    
-    // Find theme toggle in mobile menu
-    const themeToggle = page.locator('button[aria-label*="mode"]');
-    await expect(themeToggle).toBeVisible();
-    
-    // Toggle dark mode
-    await themeToggle.click();
-    await expect(page.locator('html')).toHaveClass(/theme-dark/);
-    
-    // Close mobile menu
-    if (await menuButton.isVisible()) {
-      await page.locator('[data-testid="mobile-menu-close"]').click();
-    }
-    
-    // Verify dark mode persists
-    await expect(page.locator('html')).toHaveClass(/theme-dark/);
-  });
-
-  test.skip('Should respect system preference on first visit', async ({ page }) => {
-    // Set browser to prefer dark mode
-    await page.emulateMedia({ colorScheme: 'dark' });
-    
-    await page.goto('/');
-    
-    // Should automatically be in dark mode
-    await expect(page.locator('html')).toHaveClass(/theme-dark/);
-    
-    // Set browser to prefer light mode
-    await page.emulateMedia({ colorScheme: 'light' });
-    await page.reload();
-    
-    // Should automatically be in light mode
-    await expect(page.locator('html')).not.toHaveClass(/theme-dark/);
-  });
-
-  test.skip('Theme transition should be smooth', async ({ page }) => {
-    await page.goto('/');
-    
-    // Add listener for transition events
-    const transitionPromise = page.evaluate(() => {
-      return new Promise(resolve => {
-        document.documentElement.addEventListener('transitionend', () => resolve(true), { once: true });
-        setTimeout(() => resolve(false), 1000); // Timeout after 1 second
-      });
-    });
-    
-    // Toggle theme
-    const themeToggle = page.locator('button[aria-label*="mode"]');
-    await themeToggle.click();
-    
-    // Verify transition occurred
-    const hadTransition = await transitionPromise;
-    expect(hadTransition).toBe(true);
-  });
-});
